pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "titan"
version = "0.1.0"
description = "Banking SLM fine-tuning pipeline (Qwen2.5-0.5B + LoRA, CPU)"
requires-python = ">=3.11"
dependencies = [
  "transformers>=4.45",
  "peft>=0.13",
  "datasets>=3.0",
  "accelerate>=1.0",
  "torch>=2.4",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "rouge-score>=0.1.2",
  "bert-score>=0.3.13",
  "sentence-transformers>=3.0",
  "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
